Added support basic auth for Lab3 and Lab6

+++ lab6/src/main/java/com/maxart/service/PictureResource.java
@@ -1,17 +1,26 @@
 package com.maxart.service;
 
 import com.maxart.service.exceptions.*;
+import org.postgresql.util.Base64;
 
 import java.sql.SQLException;
 import java.util.List;
+import java.util.StringTokenizer;
 import javax.ws.rs.*;
 import javax.ws.rs.core.Context;
+import javax.ws.rs.core.HttpHeaders;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.UriInfo;
 
 @Path("/pictures")
 @Produces({MediaType.APPLICATION_JSON})
 public class PictureResource {
+
+    private static final String AUTH_HEADER_KEY = "Authorization";
+    private static final String AUTH_HEADER_PREFIX = "Basic ";
+
+    private static final String USERNAME = "admin";
+    private static final String PASSWORD = "admin";
 
     @GET
     public List<Picture> find(@Context UriInfo info) {
@@ -35,8 +44,15 @@ public class PictureResource {
         return new PostgreSQLDAO().findOne(id);
     }
 
-    @POST @Consumes("application/json")
-    public String create(Picture picture) throws InvalidCreatingParametersException, InsertingException {
+    @POST
+    @Consumes("application/json")
+    public String create(@Context HttpHeaders headers, Picture picture) throws InvalidCreatingParametersException,
+            InsertingException, AuthException {
+
+        if (!isAuthenticated(headers)) {
+            throw AuthException.DEFAULT_INSTANCE;
+        }
+
         if (picture.getName() == null || picture.getName().trim().isEmpty()) {
             throw new InvalidCreatingParametersException("Invalid creating parameter: name");
         }
@@ -65,10 +81,15 @@ public class PictureResource {
         return "{\"result\":" + dao.createPicture(picture) + "}";
     }
 
-    @PUT @Consumes("application/json")
+    @PUT
+    @Consumes("application/json")
     @Path("/{id}")
-    public String update(@PathParam("id") int id, Picture picture) throws IllegalIdException, InvalidEntityException,
-            InvalidUpdatingParametersException {
+    public String update(@Context HttpHeaders headers, @PathParam("id") int id, Picture picture) throws IllegalIdException, InvalidEntityException,
+            InvalidUpdatingParametersException, AuthException {
+        if (!isAuthenticated(headers)) {
+            throw AuthException.DEFAULT_INSTANCE;
+        }
+
         if (id <= 0) {
             throw IllegalIdException.DEFAULT_INSTANCE;
         }
@@ -88,12 +109,37 @@ public class PictureResource {
 
     @DELETE
     @Path("/{id}")
-    public String delete(@PathParam("id") int id) throws IllegalIdException, InvalidEntityException {
+    public String delete(@Context HttpHeaders headers, @PathParam("id") int id) throws IllegalIdException,
+            InvalidEntityException, AuthException {
+        if (!isAuthenticated(headers)) {
+            throw AuthException.DEFAULT_INSTANCE;
+        }
+
         if (id <= 0) {
             throw IllegalIdException.DEFAULT_INSTANCE;
         }
 
         PostgreSQLDAO dao = new PostgreSQLDAO();
         return "{\"result\":" + dao.deletePicture(id) + "}";
+    }
+
+    private boolean isAuthenticated(HttpHeaders headers) {
+        List<String> authHeader = headers.getRequestHeaders().get(AUTH_HEADER_KEY);
+        if (authHeader == null) {
+            return false;
+        }
+
+        String authToken = authHeader.get(0);
+        if (authToken.isEmpty()) {
+            return false;
+        }
+
+        authToken = authToken.replaceFirst(AUTH_HEADER_PREFIX, "");
+        String decodedString = new String(Base64.decode(authToken));
+        StringTokenizer stringTokenizer = new StringTokenizer(decodedString, ":");
+        String username = stringTokenizer.nextToken();
+        String password = stringTokenizer.nextToken();
+
+        return username.equals(USERNAME) && password.equals(PASSWORD);
     }
 }
